requirements.txt
Django==5.2.6
requests>=2.32.0
python-dotenv>=1.0.1
reportlab>=3.6.13
openpyxl>=3.1.2
